Adds new alert for when IPv6 is down for too long (#836)
* Adds new alert for when IPv6 is down for too long

This commit should resolve the case where IPv6 is down for one or all
machines at a site without anyone knowing about it.

* Matches on 'site' label instead of 'machine' label

* Adds 'site' label to all platform cluster metrics

* Uses correct suffixes for bmc and v6 target rewrites

* Adds a 'site' label to all platform cluster metrics

* Removes two uses of label_replace()

* Removes stray trailing comman from alert query

* Removes yet another stray trailing comma from query

* Bumps GMX to v1.3.3

* Removes all instances of label_replace() in query

Also takes advantage new `ipv6` label in BBE metrics.

* Wraps entire query in "count by (site)"

Wrapping only the first part of the query in "count by (site)" meant
that later parts of the query could only match on the 'site' label,
meaning that any machine at a site in GMX or lame-duck mode could mask
an IPv6 issue for a different machine.

* Makes indention for all unless clauses equal