Sets strategy to Recreate for kubeip deployment
When you only have a single replica in a deployment, a rolling update
doesn't always make the most sense, especially if the service isn't user
facing. This change sets the update strategy for the kubeip deployment
to a simple "Recreate", which will just delete the old pod and create a
new one.