Updates alert for when container logs are missing
When log entries for containers on a node stop getting pushed to
Stackdriver by the Vector pod on each node, eventually the custom
log-based metric will vanish. These changes take this behavior into
account instead of looking for a log count of 0. log-based metrics in
GCP of type "Counter" do not behave the same as the "Counter" metric
type in Prometheus; they are actually more like guages:
https://cloud.google.com/logging/docs/logs-based-metrics#metric-types