Join on machine label instead of node label
The node label already exists in the metric and is inserted there by
GCP, and reflects a GKE node. I was unable to plumb the M-Lab node name
through to the node label. However, I was able to plumb a "machine"
label through, and almost all of our metrics contain both "node" and
"machine" labels with same value.