Updates alert for when Stackdriver is missing container logs for a node (#883)
* Updates github-receiver to v0.10

* Updates alert for when container logs are missing

When log entries for containers on a node stop getting pushed to
Stackdriver by the Vector pod on each node, eventually the custom
log-based metric will vanish. These changes take this behavior into
account instead of looking for a log count of 0. log-based metrics in
GCP of type "Counter" do not behave the same as the "Counter" metric
type in Prometheus; they are actually more like guages:
https://cloud.google.com/logging/docs/logs-based-metrics#metric-types

* Changes sum_over_time interval to 1h

Vector does not continually push logs, but rather waits for some
threshold or another, usually how full its local buffer is. For sites
that are not every busy, Vector may push less frequently. This commit
changes the sum_over_time() interval to 1h to be relatively sure that
Vector will have pushed at least _some_ container log entries.

* Join on machine label instead of node label

The node label already exists in the metric and is inserted there by
GCP, and reflects a GKE node. I was unable to plumb the M-Lab node name
through to the node label. However, I was able to plumb a "machine"
label through, and almost all of our metrics contain both "node" and
"machine" labels with same value.