Updates several dashboards to allow switching on site type (#925)
* Adds a new "type" field NDT:ClientRates dashboard

The new field allows the user to select between physical and virtual
sites, or both. When either is selected, then the content of the "site"
field will only contain sites that are of the selected type.

* Adds a new "Type" select box in WorkloadOverview

The new select box allows a user to select on "physical" or "virtual".

* Adds a new Type selector on SiteOverview dashboard

The contents of the existing Site selector will change depending on which Type is selected.

* Fixes regex for Site variable to work for sandbox nodes

* Renames dashboard to just NDT_Client_Rates

* Fixes capitalization in displayed dashboard name

* Auto-select a value from the 'Site' select box

* Enables multi-select for Site on SiteOverview

* Revert "Enables multi-select for Site on SiteOverview"

This reverts commit a38ee085acf763e5959c7ad64da3c5bf8850356e.

* Only shows nodes that exist at site on SiteOverview

* Causes SiteOverview to properly display site type

* Renames "Machine type" panel to "Site type"

Also, modifies the query to only display the value from one sample node
at the site.