Adds a new GKE Datasource variable to SiteOverview
The dashboard has a few panels that make use of metrics from the
prometheus-federation cluster. Previously it was using a static
datasource UID, which was only valid for the prometheus-federation
cluster in mlab-sandbox. With this commit, we can properly select the
right datasource at the top and all panels should work as expected in
all projects.