Adds a prometheus-federation "bootstrap" script
The default Cloud Build service account needs to have certain RBAC
permissions applied to it in the cluster before it can successfully
deploy the cluster completely. As sad as it is, this commit introduces
yet another small shell script which will need to be run before Cloud
Build can successfully deploy the cluster. Presumably, this script would
get run just after the cluster is created, and should only need to be
run once.

This commit also updates the README to include instructions for running
this script. vscode's markdown auto-formatter also made a number of
other style changes to the README.