Makes consistent use of $ and $$ in build config
If a variable in the build config begins with a single `$`, then it is a
built-in Cloud Build substitution variable. If a variable begins with a
double `$$`, then it will not be substituted by Cloud Build and Cloud
Build will remove the leading `$` such that bash will substitute the
variable as usual at run time.