Adds a Cloud Build configuration file (#938)
* First steps in porting build to Cloud Build

So far the Cloud Build config only sets up the basic environment and
runs apply-global-prometheus.sh. Some RBAC permissions were required:

https://cloud.google.com/kubernetes-engine/docs/how-to/role-based-access-control#permission_to_create_or_update_roles_and_role_bindings

Additionally, I added the "cloud-kubernetes-deployer" role to the
mlab-sandbox Cloud Build service account. In addition to all this, I had
to add these permissions to the role:

* container.validatingWebhookConfigurations.update
* container.mutatingWebhookConfigurations.update

* Adds additional build pieces to Cloud Build

Also, adds another Secret to the environment.

* Removes echoing of shell commands

Cloud Build does not have the same sense/ability as Travis-CI to not
echo sensitive variables to the build logs. For this reason, rather than
turning echoing on and off and on and off and on and off, etc. in the
file, we just turn off echoing altogether.

* Uses --dry-run=client instead of deprecated --dry-run

Also, installs apache2-utils to get required utility htpasswd.

* Don't echo the Linode IPv6 private ssh key

Also:
* Uses --dry-run=client in a few more places
* Uses the -m flag to gsutil when upload Prometheus targets

* Changes --dry-run client to --dry-run="client" everywhere

* Fixes one more --dry-run=client instance

* Removes static values in favor of variable

Previously, too many values were statically configured, which were only
valid for mlab-sandbox. These changes should make the script portable
across projects.

* Adds a prometheus-federation "bootstrap" script

The default Cloud Build service account needs to have certain RBAC
permissions applied to it in the cluster before it can successfully
deploy the cluster completely. As sad as it is, this commit introduces
yet another small shell script which will need to be run before Cloud
Build can successfully deploy the cluster. Presumably, this script would
get run just after the cluster is created, and should only need to be
run once.

This commit also updates the README to include instructions for running
this script. vscode's markdown auto-formatter also made a number of
other style changes to the README.

* Adds new prometheus-federation "bootstrap" script

Also:

* Replaces a static reference to "mlab-sandbox" with $PROJECT
* Fixes a grammar issue in a comment in the README

* Removes evaluated template file after application to cluster

* Makes consistent use of $ and $$ in build config

If a variable in the build config begins with a single `$`, then it is a
built-in Cloud Build substitution variable. If a variable begins with a
double `$$`, then it will not be substituted by Cloud Build and Cloud
Build will remove the leading `$` such that bash will substitute the
variable as usual at run time.

* Runs jsonlint-php on all JSON files in repo

* Removes Travis-CI config file .travis.yml

* Removes 'travis' submodule

* Adds notes to README about updating secrets in GCP