Use a custom image location for stackdriver_exporter (#944)
* Uses temp custom image for stackdriver_exporter

https://github.com/m-lab/prometheus-support/issues/896

This commit updates the stackdriver_exporter image path to a temporary
location in the measurementlab Docker hub org. Once a new version of
stackdriver_exporter is released, we should revert the image path to the
official one.

* Uses temp image location for mlabns-stackdriver_exporter

https://github.com/m-lab/prometheus-support/issues/896

This commit updates the mlabns-stackdriver_exporter image path to a
temporary location in the measurementlab Docker hub org. Once a new
version of stackdriver_exporter is released, we should revert the image
path to the official one.