Removes resource definitions for prometheus (#948)
* Configures only Prom to run on prometheus-node=true

The current prometheus-pool-big node pool has 2 fairly beefy nodes, only
one of which Prometheus actually runs on. There is no need for
stackdriver_exporter and kube-state-metrics to run on a node provisioned
for Prometheus. Let the cluster decided where to schedule other
workloads.

* Changes node pool name in kubeip

* Lowers resource requests/limits for Prometheus

The resource requests for Prometheus in mlab-sandbox were exceedinly too
high. In practice, Prometheus itself maybe uses a few CPU% and 2G of
memory (in sandbox). This lowers the limits so as to not waste a lot of
GCP resources for nothing.

* Adds toleration to Prom to run on prom node

* Increases resources limits for Prom

* Removes resource allocations from Prom and sidecars

Just let the containers use what they need. If the VM doesn't have
enough memory for the pod, then we can increase the size of the VM. This
is not a multi-tenant situation where we need to be sure that one
workload doesn't impinge on another.

* Removes resource variables for Prometheus

Just let Prometheus use as much memory as it wants, rather than
artificially constraining which sort of machine the scheduler will try
to assign the Prom pod on.