Adds new "terse" site overview dashboard (#951)
* Updates Grafana to v9.1.5

* Adds new dasbhaord Site Overview Terse

This new dashboard provides a brief summary of resource usage for each
site selected. It is suitable for viewing multiple sites at once. For
example, to view basic resource usage for all virtual sites.

* Sums network metrics instead of averaging them

For several of the machine-level metrics this dashboard averages to get
a general idea for how those resources look across all machines at the
site. However, the network metrics should be summed to display roughly
the throughput through the switch for the entire site.

* Renames Load1 panel to 'Load average (1 min)'