Promotes "GCP Billing" dashboard to GA, with some modifications (#955)
* Adds new dashboard to monitor GCP billing

This dashboard, as it is, was created by @stephen-soltesz. It was
located in Stephen's personal folder in Grafana in mlab-sandbox. This
commit is using his work as the basis for an official dashboard to
monitor GCP billing, where additional panels and metrics may be added in
the future.

* Adds new panel to estimate egress GCE network cost

Additionally:

* resizes numerous panels
* removes legend from some panels
* removes "stacked" graph display from most panels
* renames panels slightly
* removes any commens from BigQuery queries
* various other tweaks I can't think of now

* Restores "stacked" display panels for "All Projects"

And a few other minor tweaks.

* Adds new query for "Network Internet Egress%" skus

Queries the exported GCP billing data in BigQuery for any SKU like
"Network Internet Egress%". This will be grossly higher than the
Internet egress costs for just the virtual GCE platform nodes, since it
includes everything originating from GCE, not just platform nodes. But
should still provide a good idea of GCE Internet egress costs, of which
virtual platform nodes will surely be a major contributor.

* Adds description for GCE egress costs panel

* Adds new BigQuery query listing egress SKUs

It lists only the egress SKUs where the cost per day exceeds $10, which
is an arbitrary value, but weeds out quite a lot of SKUs where the cost
is mere pennies per day, and make the chart more readable.

Also, sets the cost per service per house data type to currency->USD.