Fixes and improvements to SiteOverview dashboard (#958)
* Fixes and improvements to SiteOverview dashboard

This commit started as nothing more than looking into fixing the
"Memory" panel, which was showing something fairly off of what `top` on
the machine was showing. This was fixed, but while in the dashboard I
realized that a few other things were broken or needed updating:

* The "pod" select box had no values. Fixed.
* I migrated all legacy "Graph (old)" panels to the new "Timeseries"
  type.
* Limited the "Pod CPU Usage" panel to only display timeseries where the
  value is greater than 1%.
* Set hover tip on various panels to sort descending.
* Updatesj "go routine counts" panel to have hover tip display
  "<container> (<pod>)" instead of "<deployment> <container>".
* Removes "- RSS" from "Go RSS" panel hover tip items (redundant).
* Changed the "pod" variable custom "All" value from ".*" to ".+".

* Replaces Load1 panel with a CPU panel

Load average requires too much context to understand at a glance whether
any given value is important or not. On the other hand CPU usage (as a
percentage of total) gives you a good idea of how busy the CPUs are at a
glance.

* Small improvements to SiteOverview dashboard

* Removes legend from most all panels (wasn't useful, took up space).
* Resizes "Pod" panels to same size.
* Migrates "Disk I/O" panel to newer "Timeseries" panel type.
* Moves "Process count" panel to far bottom right.
* Slight tweaks to a couple of panel names.

* Replaces Load1 panel with CPU usage panel

This commit makes the same changes to this dashboard that are being made
to the full SiteOverview dashboard. Namely:

* Replaces Load1 panel with CPU usage panel
* Updates memory panel with better representation of actual memory usage
  by adding new metrics in aggregation. Combines the "Buffer" and
  "Cache" queries into a single "Buff/Cache" query. Renames "Active"
  query to "Usage".
* Removes legend from all panels. Space is very limited on this
  dashboard, and a user can hover to see what lines represent what.