Adds OOM alerts and new OOM panel on SRE Overview dashboard (#972)
* Adds new alert for OOM events on machines

Additionally, replaces the "Kernel logs priority" panel on the SRE
Overview dashboard with OOM events. The former was never really useful,
and OOM events are much more meaningful for that space in the dashboard.

* Removes rate calculation from OOM events

GCP log-based metrics are 1m samples with a value that represents the
delta from the previous minute's value. We can use the value as is
without any calculations.

* Uses sum_over_time for OOM alert

* Removes specific numbers from comments

This way we don't have to keep the comments in sync with the actual
expression.