Updates name of node pool for kube-ip
We had a disk pressure issue on the static-public-ip nodes, which caused
script-exporter to be evicted. It is not possible to resize the disks of
an existing node pool. The only way to achieve this is to create a new
node pool with the desired configurations and then migrate your
workloads to this new node pool. I created a new node pool in mlab-oti
named static-external-ip and migrated the workloads to this new node
pool. This change configures kube-ip to be sure that those nodes get
static IPv4 addresses. Meanwhile I have manually updated the kube-ip
deployment in production with the change in this commit.