Turn off debug logging for script-exporter (#974)
* Disabled debug logging for script-exporter

Over the course of a month or so, the logs start to fill up the
filesystem, since the script appends the entire output of each script
run to a single file for each machine.

* Sets value of LOGX_DEBUG to an empty string

The script doesn't check for true or false, but just [[ -n $LOGX_DEBUG
]]. I could remove the environment variable altogether, but I sort of
like leaving it so that people down the road have visibility that this
feature even exists. It is explicity, not implicit.

* Removes env variable LOGX_DEBUG altogether from script-exporter