Fixes the K8s_APICluster dashboard
The API nodes were all renamed from "master.*" to "api.*" and the
dashboard was looking for labels with the old node name values.

I updated the dashboard title to replace the word "master", but left the
filename as is, since changing it would entail deleting the dashboard
with the new name and creating a new one. Notwithstanding the file name,
all user-visible instances of "master" should now be gone.