Changes downloader alert to fire after 72h, not 21h (#1035)
The downloader used to download databases about every 8h. We have reconfigured
th downloader to only download about every 24h. Additionally, there is no need
for this alert to be super sensitive. It is okay if annotations happen with
databases that are only a couple days old. However, if downloader can't
download all the files for more than 3d, then fire an alert.