Changes all BigQuery datasource types of doitintl to grafana
The doitintl version of the Grafana BigQuery datasource plugin is deprecated
(or even removed) in favor of the native version maintained by Grafana. A while
we (I?) had migrated all dashboards to use the Grafana maintained plugin, I
think, but there were still a few dashboard referencing it for some reason.
This commit removes those references and replaces them with references to the
Grafana BQ plugin.

Additionally this commit outright removes the templates for the doitintl BQ
datasource plugin.