Fully removes doitintl BigQuery datasource references (#1049)
* Updates SwitchOverview dashboard to use lastest graph type

The dashboard was using the old, deprecated "Graph (old)" panel types, which
were deprecated a long time ago in favor of the newer, more performant "Time
Series" panel type. I simply clicked the "Migrate" button in each of the
panels, and Grafana seems to have handled the rest.

* Updates Grafana to v11.1.0

* Updates Prometheus to v2.53.1

* Adds old doitintl BQ plugins to list of datasources to delete

These were replaced by a native Grafana BigQuery datasource plugin.

* Changes all BigQuery datasource types of doitintl to grafana

The doitintl version of the Grafana BigQuery datasource plugin is deprecated
(or even removed) in favor of the native version maintained by Grafana. A while
we (I?) had migrated all dashboards to use the Grafana maintained plugin, I
think, but there were still a few dashboard referencing it for some reason.
This commit removes those references and replaces them with references to the
Grafana BQ plugin.

Additionally this commit outright removes the templates for the doitintl BQ
datasource plugin.

* Adds new prune:true config to auto delete provisioned datasources

As of this version (v11.1.x) Grafana apparently supports automatically removing
provisioned datasources when the provisioning source for the datasource no
longer exists. This should hopefully obviate the need for the deleteDatasources
configuration.

* Adds back deleteDatasources, but only with existing datasources

No need to keep already-deleted datasources in the array.

* Adds manageAlerts:false to all Prometheus datasources

It's possible this setting is true by default, which seems like it might be
causing Grafana to query the API for a list of rules, which returns a 404,
which in turn causes an obscure "Not found" popup in the Grafana display.

* Temporarily enable login form

* Temporarily disable auth.proxy for Grafana

* Re-disabled login form, and re-enables auth.proxy

* Removes deprecated legacy [alerting] option

* Bumps Grafana to latest patch version v11.1.3

* Removes manageAlerts:false from all datasources

I had added this configuration in the hopes that it would clear up a 404 error
that Grafana was hitting on the Home page, but it didn't work. It turns out
what I was seeing was a bug, which was correct:

https://github.com/grafana/grafana/pull/89905