Removes script_success metrics from alerts + deletes 2 dashboards
We are going to retire script_exporter, and along with e2e testing of ndt5.
This commit removes the script_success metrics from any alerts in this
repository. It also deletes one dashboard related to script_exporter.

While I was at it, I deleted a defunct MLABNS dashboard that displays no data
and is probably no longer useful.