Increases script-exporter timeout for wehe from 50s to 55s
Some manual testing in sandbox was showing that the wehe e2e tests were taking
about 51.25s to complete (successfully), but when run by script-exporter it was
getting killed by the timeout. In production the wehe e2e tests seem to
*mostly* complete in 15 to 20 seconds, though there are some completion times
in production >40s. This commit just gives the wehe client a few more seconds
to complete.