Changes wehe e2e test from using applemusic to amazon
I noticed that many e2e tests in sandbox were timing out at 50s from the
timeout setting configured for the wehe_client script. But in production the
tests were completing in around 10 or 15 seconds.

I turned on debug logging for wehe-cmdline.jar and noticed that for some odd
reason there was a 20s delay when sending packet number 13 our of 15. And this
happens twice during the test, making the test very long.

Looking at the wehe-cmdline repository, I noticed that a new version was
released just a couple months ago which updated all the replay files.
Inspecting the applemusic replay file (which is JSON), I noticed that the
timestamp field of the 13th packet in the array has a value of ~20s, which I
think explain why there is always a 20s delay when sending the 13th packet.

I looked through the available replay files looking for one where the final
packet's timestamp value was low, and it turns out that the amazon test fit
that bill. This should cause wehe e2e tests to complete much more quickly.