Moves autonode e2e monitoring to the autojoin cluster
I _believe_ that Stephen implemented the e2e script-exporter monitoring of
autonodes in the prometheus-federation cluster because the Locate Services
queries this Prometheus instance for script_success metrics to determine
whether to server a particular host. However, it make more sense to do all
Autojoin monitoring in the autojoin cluster, and then simply do a federations
scrape of those script_success metrics. This separates concerns better.