Adds new "Port Status" panel to Autojoin:SiteOverview
The "OR vector(999)" part of the query for the panel is to aid in Value
Mappings for the panel. The main expression could return a range of values, all
of which indicate some problem, however it is highly unlikely to ever return
999. This allows the panel to display "OK" when the value is 999, and "FAIL"
when the value is anything between the arbitrary range of 0-100.