Improve monitoring for the autojoin cluster (#1077)
* Adds new alert for when the Autojoin API is returning too many errors

Actually, most of the changes in this PR are to even support alerts from the
autojoin cluster.

* Gets Alertmanage basic auth k8s Secret from GCP Secret Manager

* Adds missing comma to autojoin prometheus manifest

* Mounts the alertmanager credentials volume in correct container

I had accidentally mounted it in the config-reload container. This moves it to
the prometheus container.

* Fixes Alertmanager URL in Prometheus configuration

* Moves autonode e2e monitoring to the autojoin cluster

I _believe_ that Stephen implemented the e2e script-exporter monitoring of
autonodes in the prometheus-federation cluster because the Locate Services
queries this Prometheus instance for script_success metrics to determine
whether to server a particular host. However, it make more sense to do all
Autojoin monitoring in the autojoin cluster, and then simply do a federations
scrape of those script_success metrics. This separates concerns better.

* Fixes a typo in the autojoin federation scrape job

* Subs the AUTOJOIN_PROJECT var in Prometheus config file

* Fixes byos-nodes script job in the autojoin cluster

* Puts the AUTOJOIN_PROJECT template sub in right Prom config file

* Changes substitution variable name for federation scrape of autojoin

* Remove AUTOJOIN_CLUSTER template variable substitution

* Removes gkedatasource variable from Autojoin:SiteOverview dashboard

The e2e script-exporter metrics are now coming from the autojoin cluster.

* Adds blackbox_exporter probing for sidecars and ndt-server in autojoin

* Apply the blackbox ConfigMap instead of replace

The replace operation fails when the ConfigMap doesn't already exist. I can't
recall why I set it to replace some years back. Let's try with apply.

* Mounts blackbox-target in prometheus-storage

* Fix ndt-server blackbox target port

It was missing a colon between the hostname and port. Also, for sidecars, the
target will already have a port, so there is no need for the regex and
reconstruction of the host:port, just use the whole target as is.

* Moved the blackbox Service manifest to the right directory

* Adds new "Port Status" panel to Autojoin:SiteOverview

The "OR vector(999)" part of the query for the panel is to aid in Value
Mappings for the panel. The main expression could return a range of values, all
of which indicate some problem, however it is highly unlikely to ever return
999. This allows the panel to display "OK" when the value is 999, and "FAIL"
when the value is anything between the arbitrary range of 0-100.

* Remove generation of script-targets in prometheus-federation

e2e script-exporter monitoring of autonodes now happens in the autojoin
cluster.

* Adds cluster=autojoin to federated scrape of the autojoin cluster

Also add cluster=data-pipeline for same scrape of data-pipeline

* Adds 2 new alerts for the Autojoin API

* Uses new -client-name flag for ndt7 e2e monitoring of autonodes

* Adds a bigquery_exporter deployment to the autojoin cluster

Also adds one query that calculates the number of Autojoin ndt7 rows in the
last 24h, offset by 4h.

* Makes new autojoin bq_exporter query a template

The BigQuery project needs to be changed on a project-by-project basis.

* Adds new alert for ndt7 test to BigQuery row skew

* Updates script_exporter image to v0.2.13 in autojoin cluster

* Updates BigQuery rows/ndt7 test alert

* Fixes to the Autojoin:SiteOverview dashboard