Fixes a small bug in the Autojoin_SiteOverview dashboard
The panels were not correctly filtering on org name, so if for any reason the
same machine name belonged to two orgs, then the panel would not display
properly. This shouldn't normally happen, but, for example, I accidentally
created an org with the wrong name, the ISP brought up the machine, then I
realized my mistake and created a new org with the right name, and the ISP
brought that one up. We then had metrics for the same machine name, but two
different orgs. For some reason this bug is affecting the "mlab" org too, and
this should fix that also.