Adds a new BYOS panel to the NDT Global Test Rate dashboard
It may be possible to combine regular NDT query results with BYOS query results
using a Grafana transform of some sort, but for now this change adds BYOS test
rate metrics in its own distinct panel. This should be plenty good for now, and
probably makes it easier to get a sense of how much traffic is coming from each
deployment type.