Add null Alertmanager receiver to stop FYI "alerts" from hitting default receiver (#1092)
* Removes default receiver from routes section

Having the default receiver in the routes block was redundant. I have also
added a new matcher to the default receiver to be sure that FYI alerts never
fall through to the default receiver

* Adds new "null" receiver that takes no action

This new receiver will be used to avoid FYI "alerts" from passing on to the
default receiver, which is Slack.

* Adds missing colon after map element