Disabled the AutojoinDNSRecordTooCloseToExpiring alert (#1093)
* Disabled the AutojoinDNSRecordTooCloseToExpiring alert

This alert was implemented with the notion that M-Lab-managed VMs would be
leveraging the Autojoin system, but that never happened. However, I have left
the alert commented out, just in case this changes in the future.

* Removes the AutojoinDNSRecordTooCloseToExpiring alert

The alert isn't used for anything, but is causing unactionable alerts in
sandbox and staging.